requirements.txt
absl-py==2.1.0
accelerate==1.4.0
ale-py==0.10.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.1.0
Babel==2.15.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==1.2.0
boto3==1.35.44
botocore==1.35.44
Brotli==1.1.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.3
decorator==5.1.1
deeplake==3.9.26
defusedxml==0.7.1
dill==0.3.9
distlib==0.3.9
dm-tree==0.1.8
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
etils==1.10.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.5
fastapi-cli==0.0.5
fastjsonschema==2.20.0
filelock==3.16.1
flake8==7.0.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
gensim==3.8.0
gevent==24.11.1
geventhttpclient==2.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.65.5
gymnasium==1.1.1
h11==0.14.0
h5py==3.11.0
holidays==0.74
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
hub==3.0.1
huggingface-hub==0.29.0
humbug==0.3.2
idna==3.6
immutabledict==4.2.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
locust==2.32.3
lz4==4.3.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.1.0
multiprocess==0.70.17
murmurhash==1.0.12
namex==0.0.8
narwhals==1.27.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.12
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
prophet==1.1.7
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
PyMuPDF==1.25.3
PyOpenGL==3.1.9
pyparsing==3.1.2
pypdfium2==4.30.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.10.3
safetensors==0.5.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==69.2.0
shellingham==1.5.4
simple-parsing==0.1.6
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.5.1
stack-data==0.6.3
stanio==0.5.1
starlette==0.41.3
streamlit==1.42.1
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-cpu==2.18.0
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.49.0
Tree==0.2.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
ultralytics==8.3.151
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.32.1
virtualenv==20.29.3
wasabi==1.1.3
watchdog==6.0.0
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.44.0
widgetsnbextension==4.0.11
wordcloud==1.9.4
wrapt==1.16.0
WTForms==3.1.2
zipp==3.20.2
zope.event==5.0
zope.interface==7.2
